fixed type issue

+++ lunabot_embedded/firmware/teensy_main/lib/lunabot_drivers/robot.hpp
@@ -21,13 +21,13 @@ extern Sabertooth MC3; // bottom
 
 namespace actuation {
 void cb(int8_t lin_act);
-void update(int32_t &);
+void update(float &);
 
 } // namespace actuation
 
 namespace drivetrain {
 void cb(int8_t left, int8_t right);
-void update(int32_t &, int32_t &, float &, float &);
+void update(float &, float &, float &, float &);
 float update_curr_left();
 float update_curr_right();
 } // namespace drivetrain
@@ -38,13 +38,13 @@ void update(float &d0, float &d1, float &d2);
 
 namespace deposition {
 void cb(int8_t dep);
-void update(int32_t &);
+void update(float &);
 float update_curr();
 } // namespace deposition
 
 namespace excavation {
 void cb(int8_t exc);
-void update(int32_t &, float &);
+void update(float &, float &);
 float update_curr();
 } // namespace excavation
 #endif
